Fix sign compare issue.

+++ src/shape.cpp
@@ -78,7 +78,7 @@ namespace picovector {
   // }
 
   void shape::stroke(float thickness) {
-    for(int i = 0; i < this->paths.size(); i++) {
+    for(int i = 0; i < (int)this->paths.size(); i++) {
       this->paths[i].stroke(thickness);
     }
   }
